refactored Navbar and fixed logo sizing
