+++ src/assets/Navbar.js
@@ -1,24 +1,13 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import { retrieveFullSearchResults } from '../store/actions';
-import { Link } from 'react-router-dom';
-import { Menu, Row, Col, Avatar, Dropdown, Input } from 'antd';
-import {
-  UserOutlined,
-  QuestionOutlined,
-  SettingOutlined,
-  ShopOutlined,
-  FileOutlined,
-} from '@ant-design/icons';
+import { Row, Col } from 'antd';
+
 import AlumniLogo from '../img/AlumniLogo.svg';
+import ProfileIcon from './ProfileIcon';
+import SearchBar from './SearchBar';
 
 const Navbar = (props) => {
-  const handleSearchChange = (e) => {
-    if (e.target.value !== '') {
-      props.retrieveFullSearchResults(e.target.value);
-    }
-  };
-
   return (
     <Row
       style={{
@@ -30,7 +19,7 @@ const Navbar = (props) => {
     >
       <Col span={6} style={{ display: 'flex', justifyContent: 'flex-start' }}>
         <div
-          style={{ width: '60%' }}
+          style={{ width: '60%', maxWidth: '200px' }}
           onClick={() => {
             props.history.push('/');
           }}
@@ -46,36 +35,7 @@ const Navbar = (props) => {
           justifyContent: 'center',
         }}
       >
-        <Dropdown
-          placement="bottomCenter"
-          trigger={['click']}
-          overlay={
-            <Menu>
-              {props.searchResults.posts.map((p) => (
-                <Menu.Item icon={<FileOutlined />} key={p.id}>
-                  <Link to={`post/${p.id}`}>{p.title}</Link>
-                </Menu.Item>
-              ))}
-              {props.searchResults.users.map((u) => (
-                <Menu.Item icon={<UserOutlined />} key={u.id}>
-                  <Link to={`user/${u.id}`}>{u.display_name}</Link>
-                </Menu.Item>
-              ))}
-              {props.searchResults.rooms.map((r) => (
-                <Menu.Item icon={<ShopOutlined />} key={r.id}>
-                  <Link to={`room/${r.id}/page/1`}>{r.room_name}</Link>
-                </Menu.Item>
-              ))}
-            </Menu>
-          }
-        >
-          <Input.Search
-            placeholder="input search text"
-            onChange={handleSearchChange}
-            onSearch={(e) => console.log(e)}
-            enterButton
-          />
-        </Dropdown>
+        <SearchBar/>
       </Col>
       <Col
         span={6}
@@ -85,26 +45,7 @@ const Navbar = (props) => {
           alignItems: 'center',
         }}
       >
-        <Dropdown
-          placement="bottomRight"
-          arrow
-          trigger={['click']}
-          overlay={
-            <Menu>
-              <Menu.Item icon={<QuestionOutlined />}>
-                <Link to="/faq">FAQ</Link>
-              </Menu.Item>
-              <Menu.Item icon={<UserOutlined />}>
-                <Link to={`/user/${props.user.id}`}>My Profile</Link>
-              </Menu.Item>
-              <Menu.Item icon={<SettingOutlined />}>
-                <Link to="/settings">Settings</Link>
-              </Menu.Item>
-            </Menu>
-          }
-        >
-          <Avatar src={props.user.profilePicture} size="large" />
-        </Dropdown>
+        <ProfileIcon />
       </Col>
     </Row>
   );
